Matter IDL compatibility checker as a part of matter-idl
