+++ .github/workflows/protocol_compatibility.yaml
@@ -34,7 +34,6 @@ jobs:
        - name: Install dependencies
          run: |
            python3 -m venv venv
-           venv/bin/pip3 install click coloredlogs lark
            venv/bin/pip3 install -e scripts/py_matter_idl
        - name: Create old/new copies
          run: |
@@ -46,4 +45,4 @@ jobs:
            patch -p1 <out/patch.diff
        - name: Check backwards compatibility
          run: |
-           venv/bin/python3 scripts/backwards_compatibility_checker.py out/old_version.matter out/new_version.matter
+           venv/bin/matter-idl-check-backward-compatibility out/old_version.matter out/new_version.matter
