+++ scripts/py_matter_idl/setup.cfg
@@ -25,6 +25,8 @@ classifiers =
 packages = find_namespace:
 include_package_data = True
 install_requires =
+    click
+    coloredlogs
     lark
     jinja2
 
@@ -33,6 +35,7 @@ console_scripts =
     matter-data-model-xml-parser = matter.idl.data_model_xml:main
     matter-idl-lint-parser = matter.idl.lint:main
     matter-idl-parser = matter.idl.matter_idl_parser:main
+    matter-idl-check-backward-compatibility = matter.idl.backwards_compatibility:main
     matter-zapxml-parser = matter.idl.zapxml:main
 
 [options.packages.find]
